change the names of signUpComp to begin and end
