+++ src/components/auth/signUpComps/beginComp/progressBegin.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
-export default function Progress1() {
+export default function ProgressBegin() {
   return (
-    <React.Fragment>
+    <>
       <img style={{ marginBottom: "18px" }} src={"/images/prog1.png"} alt="progress1" width={"100%"} height="12px"/>
-    </React.Fragment>
+    </>
   )
 }
