--- src/components/auth/signUpComps/comp1/signUpComp1.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import Progress1 from './progress1'
-import SignUpTitle from '../signUpTitle'
-import InputsComp1 from './inputsComp1'
-
-export default function SignUpComp1({}) {
-    
-    return (
-        <div >
-            <Progress1 />
-            <SignUpTitle />
-            <InputsComp1 />
-        </div>
-    )
-}
