--- src/components/auth/signUpComps/comp2/signUpComp2.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import SignUpTitle from '../signUpTitle'
-import InputComp2 from './inputComp2'
-import Progress2 from './progress2'
-
-export default function SignUpComp2() {
-    return (
-            <div>
-                <Progress2 />
-                <SignUpTitle />
-                <InputComp2 />
-            </div>
-    )
-}
