+++ src/components/auth/signUpComps/endComp/progressEnd.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Progress2() {
+export default function ProgressEnd() {
   return (
     <React.Fragment>
       <img style={{ marginBottom: "18px" }} src={"/images/prog2.png"} alt="progress2" width="100%" height="12px"/>
