+++ src/components/auth/signUpComps/endComp/signUpEndComp.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import SignUpTitle from '../signUpTitle'
+import ProgressEnd from './progressEnd'
+import InputsEnd from './inputsEnd'
+
+export default function SignUpEndComp() {
+    return (
+            <>
+                <ProgressEnd />
+                <SignUpTitle />
+                <InputsEnd />
+            </>
+    )
+}
