+++ src/components/auth/signUpComps/imagesPosts.jsx
@@ -2,10 +2,8 @@ import React from 'react'
 
 export default function ImagesPosts(props) {
   return (
-
-    <div >
+    <>
       <img src="/images/imagesPosts1.png" alt="imagesPosts" width={props.width} />
-    </div>
-
+    </>
   )
 }
