+++ src/components/auth/signUpComps/signUpPage.jsx
@@ -1,37 +1,30 @@
 import React, { useState } from 'react'
 import Logo from '../../general_comps/logo'
-import SignUpComp1 from './comp1/signUpComp1'
-import SignUpComp2 from './comp2/signUpComp2'
+import SignUpBeginComp from './beginComp/signUpBeginComp'
+import SignUpEndComp from './endComp/signUpEndComp'
 import ImagesPosts from './imagesPosts'
 import { useSelector } from "react-redux";
 import { Link } from 'react-router-dom'
 
 
-
 export default function SignUpPage() {
-  const { showHideComp1, showHideComp2, form } = useSelector(myStore => myStore.signUpSlice)
-
-  // console.log(form)
+  const { isShowBeginComp, isShowEndComp, form } = useSelector(myStore => myStore.signUpSlice)
 
   return (
-    <div>
       <div className='d-flex justify-content-lg-between justify-content-center'>
         <div className='col-11 col-sm-8 col-md-6 col-lg-5 col-xl-4'>
           <div className='ps-lg-5 ms-lg-5 me-lg-1'>
 
             <div className='d-flex justify-content-between mb-3 mb-lg-0'>
               <Link to="/"> <Logo margin="mb-5" /></Link>
-              {/* <div className='d-lg-none d-block col-5 mt-5 mb-3 '>
-                <ImagesPosts width={"100%"} />
-              </div> */}
             </div>
 
-            {showHideComp1 &&
-              <SignUpComp1 />
+            {isShowBeginComp &&
+              <SignUpBeginComp />
             }
 
-            {showHideComp2 &&
-              <SignUpComp2 />
+            {isShowEndComp &&
+              <SignUpEndComp />
             }
 
           </div>
@@ -42,7 +35,5 @@ export default function SignUpPage() {
         </div>
 
       </div>
-    </ div >
-
   )
 }
