+++ src/components/auth/signUpComps/signUpTitle.jsx
@@ -2,10 +2,9 @@ import React from 'react'
 
 export default function SignUpTitle() {
     return (
-        <div >
+        <div>
             <h2 className='s30 dark'>Sign Up</h2>
             <h6 style={{ marginBottom: "25px"}} className='s14 gray weight400'>Create account to start using HealthyMe.</h6>
-       
         </div>
     )
 }
